First version of app done

+++ QuizGame/GameSession/GameSession.swift
@@ -9,8 +9,8 @@ import UIKit
 
 final class GameSession: UIViewController {
 
-    var result = 0
-    var questionsCount = 0
+    var result = 0.0
+    var questionsCount = 0.0
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -24,7 +24,7 @@ final class GameSession: UIViewController {
 
 extension GameSession: GameControllerDelegate {
 
-    func gameDidEndWith(_ result: Int, questionsCount: Int) {
+    func gameDidEndWith(_ result: Double, questionsCount: Double) {
 
         self.result = result
         self.questionsCount = questionsCount
